refactor: normalize amount and available_amount to int across architectures (#27)
* refactor: normalize amount and available_amount to int across architectures

* feat: add numconv utility for int-to-int32 conversion and integrate it across services for safer type handling

* refactor(item-service): extract item aggregation logic into reusable function

* docs(database-schema): update `available_amount` type from BIGINT to INT

* fix(transaction-service): handle invalid input errors in gRPC item client

* fix(transaction-service): handle proto mapping errors and add overflow test cases

* test(api-gateway): add conditional test case for `available_amount` overflow on 64-bit systems

* test(api-gateway): add 64-bit specific test case for `amount` overflow in transaction client

+++ microservices/api-gateway/internal/dto/item.go
@@ -5,7 +5,7 @@ package dto
 type Item struct {
 	ID              string `json:"id"`
 	Name            string `json:"name"`
-	AvailableAmount int64  `json:"available_amount"`
+	AvailableAmount int    `json:"available_amount"`
 	CreatedAt       string `json:"created_at"`
 	UpdatedAt       string `json:"updated_at"`
 }
@@ -23,7 +23,7 @@ type SyncItemsRequest struct {
 type SyncItemInput struct {
 	ID              *string `json:"id,omitempty"`
 	Name            string  `json:"name"`
-	AvailableAmount int64   `json:"available_amount"`
+	AvailableAmount int     `json:"available_amount"`
 }
 
 type PaginationMeta struct {
